+++ packages/babel-plugin-transform-lwc-class/src/component.js
@@ -28,11 +28,6 @@ module.exports = function ({ types: t, }) {
 
                 const classBody = path.get('body');
 
-                // Deal with component labels
-                const labels = getComponentLabels(classBody);
-                const existingLabels = state.file.metadata.labels || [];
-                state.file.metadata.labels = [...existingLabels, ...labels];
-
                 if (isDefaultExport(path)) {
                     const declaration = path.parentPath.node;
                     if (declaration.leadingComments) {
@@ -78,35 +73,6 @@ module.exports = function ({ types: t, }) {
     function importDefaultTemplate(state) {
         const componentName = getBaseName(state);
         return state.file.addImport(`./${componentName}.html`, 'default', 'tmpl');
-    }
-
-    function getComponentLabels(classBody) {
-        const labels = [];
-
-        const labelProperty = findClassProperty(classBody, LWC_COMPONENT_PROPERTIES.LABELS, { static: true });
-        if (labelProperty) {
-            if (!labelProperty.get('value').isArrayExpression()) {
-                throw labelProperty.buildCodeFrameError(
-                    `"labels" static class property should be an array of string`
-                );
-            }
-
-            labels.push(
-                ...labelProperty.get('value.elements').map(labelValue => {
-                    if (!labelValue.isStringLiteral()) {
-                        throw labelValue.buildCodeFrameError(
-                            `Label is expected to a be string, but found ${labelValue.type}`
-                        );
-                    }
-
-                    return labelValue.node.value;
-                })
-            );
-
-            labelProperty.remove();
-        }
-
-        return labels;
     }
 
     function wireTemplateToClass(state, classBody) {