+++ packages/babel-plugin-transform-lwc-class/src/utils.js
@@ -10,16 +10,6 @@ function findClassMethod(path, name, properties = {}) {
     ));
 }
 
-function findClassProperty(path, name, properties = {}) {
-    path.assertClassBody();
-
-    return path.get('body').find(path => (
-        path.isClassProperty({ static: properties.static }) &&
-        path.get('key').isIdentifier({ name })
-    ));
-}
-
-
 function isClassMethod(classMethod, properties = {}) {
     const { kind = 'method', name } = properties;
     return classMethod.isClassMethod({ kind }) &&
@@ -75,7 +65,6 @@ function getImportSpecifiers(path, sourceName) {
 
 module.exports = {
     findClassMethod,
-    findClassProperty,
     isClassMethod,
     isGetterClassMethod,
     isSetterClassMethod,